chore: relocate peek_next_commit onto the logstore directly for easier use
This will help retry logic associated with the logstore to directly find
the appropriate next c ommit

Related to #3306
